Library API change: lame_decode functions are now obsolete but still present, please use hip_decode instead.
The reason for this change is: lame_decode functions use a single global variable within the library to store decoder setup, hip_decode functions don't.

The encoder now uses hip_decode internally and it is now possible to use clipdetect feature while reencoding mp3 to mp3.

+++ libmp3lame/id3tag.c
@@ -28,7 +28,7 @@
  * NOTE: See http://id3.org/ for more information about ID3 tag formats.
  */
 
-/* $Id: id3tag.c,v 1.53.2.3 2008/09/14 11:51:49 robert Exp $ */
+/* $Id: id3tag.c,v 1.53.2.4 2008/10/11 18:04:05 robert Exp $ */
 
 #ifdef HAVE_CONFIG_H
 #include <config.h>
@@ -372,7 +372,7 @@ as follows.
 (although some are not exactly same)*/
 
 int
-id3tag_set_albumart(lame_global_flags * gfp, const char *image, unsigned long size)
+id3tag_set_albumart(lame_global_flags * gfp, const char *image, size_t size)
 {
     int     mimetype = 0;
     unsigned char const *data = (unsigned char const *) image;
